requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
celery==5.3.6
redis==5.0.3
google-genai==0.3.0
pydantic==2.6.4
python-multipart==0.0.9
requests==2.31.0
